requirements.txt
ccxt
pandas
python-dotenv
ta
streamlit
plotly
watchdog
scikit-learn
joblib
